fade between effects

+++ microphone.h
@@ -1,4 +1,4 @@
 bool setupMicrophone();
 void processAudioStream();
 void onAudioDataReceived();
-float getAudioIntensityRatio();
+double getAudioIntensityRatio();
